Create new file to contain methods
