--- create.rb
@@ -1,3 +1,0 @@
-require_relative './teacher'
-require_relative './student'
-require_relative './books'
